backend/requirements.txt
black==25.1.0
Django==5.2.3
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
PyJWT==2.9.0

